enabled edit and delete function on users

+++ app/Http/Controllers/AdminUsersController.php
@@ -4,13 +4,13 @@ namespace App\Http\Controllers;
 
 use Illuminate\Http\Request;
 use App\Models\User;
-use Illuminate\Support\Facades\Storage;
+use Illuminate\Support\Facades\Auth;
 
 class AdminUsersController extends Controller
 {
     public function index() {
 
-        $users = User::paginate(2);
+        $users = Auth::user()->paginate(2);
 
         return view('admin.users.index', compact('users'));
 
@@ -18,7 +18,7 @@ class AdminUsersController extends Controller
 
     public function editUser($id) {
 
-        $user = User::findOrFail($id);
+        $user = Auth::user()->findOrFail($id);
 
         return view('admin.users.edit', compact('user'));
 
@@ -42,9 +42,13 @@ class AdminUsersController extends Controller
         
         // photo
         if($file = $request->file('photo')) {
-            
+           
             //Finds the old picture and deletes it
-            unlink(public_path() . $user->photo); 
+            if($user->photo !== null && strpos($user->photo, 'lorempixel') == FALSE) {
+                
+                unlink(public_path() . $user->photo); 
+
+            }
 
             $name = time() . $file->getClientOriginalName();
 
@@ -57,6 +61,27 @@ class AdminUsersController extends Controller
         $user->update($inputs);
 
         return redirect('/admin/users')->with('user_updated', 'User has been updated');
+
+    }
+
+    public function deleteUser(User $user) {
+
+        if(Auth::user() != $user) {
+
+            //Finds the old picture and deletes it
+            if($user->photo !== null && strpos($user->photo, 'lorempixel') == FALSE) {
+                
+                unlink(public_path() . $user->photo); 
+
+            }
+
+            $user->delete();
+    
+            return redirect('/admin/users')->with('user_deleted', 'User has been deleted');
+
+        }
+            
+        return redirect('/admin/users')->with('user_deleted', 'Logged in user cannot be deleted');
 
     }
 }
